backend/requirements.txt
fastapi
uvicorn
pydantic
bcrypt
pyjwt
torch
transformers
datasets
python-telegram-bot
gtts
SpeechRecognition
pydub
requests
